requirements.txt
flask==3.0.0
PyMySQL[rsa]==1.1.0
gunicorn==20.1.0
redis==4.4.2
elasticsearch==8.11.0
SQLAlchemy==2.0.23
#Flask-SQLAlchemy==3.1.1
gevent==22.10.2
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
kafka-python==2.0.2
pandas==2.1.3
tensorflow==2.15.0
tensorflow_hub==0.15.0
